[MER-586] Catch Initialization (#20)
## CatchOptions
Adds the `CatchOptions` object to specify the optional configuration settings which control the global behavior of the Catch SDK.
### Options include: 
`theme`: Specifies the default global theme which determines the look and feel of widgets. Defaults to the "light-color" predesigned theme.
`environment`: Either sandbox (to be used for development and testing), or production (to be used in live applications).
`useCatchFonts`: Specifies if custom Catch fonts should be loaded into the application. Defaults to true.

## Catch Initialization

Adds initialization code for the Catch SDK: 
```
Catch.initialize(merchantPublicKey: testPublicKey, options: options) { result in
  switch result {
  case .success:
    // The application may handle success in this block if desired.
  case .failure(let error):
    // The application may handle failure in this block as needed.
  }
}
```
### Initialization Parameters
`publicKey`: A string representing the merchant public API key.
`options`: The CatchOptions specified above to specify global SDK behavior.
`completion`: A callback function to handle the success or failure of initialization.