Support iPad Multiple Scenes (#91)
### Summary

This PR adds support for iPads with multiple scenes in the `UIApplication` extension. With the introduction of multiple scenes in iOS 13, iPads can have multiple windows and view controllers, requiring adjustments to the existing code.

To handle iPads with multiple scenes, the `topViewController()` method in the `UIApplication` extension has been updated. The changes ensure that the topmost view controller is correctly determined, taking into account both the primary scene and additional connected scenes.